fastapi_backend/requirements.txt
# Runtime
fastapi==0.88.0
pydantic==1.10.2
uvicorn==0.20.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.21.0

# Development
black==22.10.0
flake8==6.0.0
pytest==7.2.0

## The following requirements were added by pip freeze:
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
ecdsa==0.18.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==1.7.1
pyparsing==3.0.9
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
